+++ addon/components/tt-time-picker.js
@@ -5,6 +5,8 @@ import Ember from 'ember';
 import layout from '../templates/components/tt-time-picker';
 import ClickOutsideMixin from 'ember-ui-components/mixins/click-outside';
 
+const { computed } = Ember;
+
 /**
   @class TimePickerComponent
   @namespace Time
@@ -37,23 +39,78 @@ export default Ember.Component.extend(ClickOutsideMixin, {
   output: 'date',
 
   /**
+    @property displayFormat
+    @type {String}
+    @default `hh:mm a`
+  */
+  displayFormat: 'hh:mm a',
+
+  /**
+    ## Time interval
+
+    The `timeInterval` specifies the interval used for the time picker.  An interval
+    of 10 will produce a time picker that will have times every ten minutes starting
+    on the hour.  If set to 30, the default, the time picker will have half-hour
+    intervals.
+
+    @property timeInterval
+    @type {Integer}
+    @default `30`
+  */
+  timeInterval: 30,
+
+  /**
     @property selectedTime
     @type {Object}
   */
   selectedTime: null,
 
   /**
+    This computed property takes the `selectedTime` property, which is a javascript
+    `Date` object and returns an object with `hour` and `minute` properties.
+
+    @property _selectedTime
+    @type {Object}
+    @private
+  */
+  _selectedTime: computed('selectedTime', function () {
+    let selectedTime = this.get('selectedTime');
+    if (!selectedTime) { return selectedTime; }
+    if (Ember.typeOf(selectedTime) !== 'date') {
+      selectedTime = new Date(selectedTime);
+    }
+    return { hour: selectedTime.getHours(), minute: selectedTime.getMinutes() };
+  }),
+
+  /**
+    This is an array of time objects. Each object has two properties: `hour` and
+    `minute`.  There is one object for each time slot available on the time picker.
+
+    ```
+    [
+      { hour: 0, minute: 0 },
+      { hour: 0, minute: 30 },
+      { hour: 1, minute: 0 },
+      { hour: 1, minute: 30 },
+      { hour: 2, minute: 0 },
+      // ...
+    ]
+    ```
+
     @property times
     @type {Array}
     @private
   */
-  times: Ember.computed(function() {
-    let timesArray = Ember.A();
-    for (let i = 0; i < 24; i++) {
-      timesArray.pushObject({ hour: i, minute: 0 });
-      timesArray.pushObject({ hour: i, minute: 30 });
+  times: Ember.computed('timeInterval', function() {
+    let i, d, output = [];
+    let interval = this.get('timeInterval') || 1;
+    let divisions = 60 / interval;
+    for (i = 0; i < 24; i++) {
+      for (d = 0; d < divisions; d++) {
+        output.push({ hour: i, minute: interval * d });
+      }
     }
-    return timesArray;
+    return output;
   }),
 
   /**
@@ -61,6 +118,7 @@ export default Ember.Component.extend(ClickOutsideMixin, {
     @private
   */
   handleClickOutside() {
+    if (this.get('isDestroyed') || this.get('isDestroying')) { return; }
     this.sendAction('close');
   },