+++ addon/helpers/compare-json.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export function compareJson([a, b]/*, hash*/) {
+  return JSON.stringify(a) === JSON.stringify(b);
+}
+
+export default Ember.Helper.helper(compareJson);