+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+### v0.2.0 2016-08-12
+
+* [FEATURE] Add `timeInterval` property to the TimePickerComponent.
+
+* [FEATURE] Add `displayFormat` property to the TimePickerComponent.
+
+* Add  `timeInterval` and `pickerDisplayFormat` properties to the InputTimeComponent.
+
+* Update the TimePickerComponent to refactor away from using a component for every time slot.
+
+* Remove TimeSlotComponent
+
+
+
+
 ### v0.1.0 2016-06-26
 
 * [FEATURE] Add `displayFormat` property to InputDateComponent and InputTimeComponent.